Skills maint, and starting to link the different screens together

+++ routes.go
@@ -30,6 +30,7 @@ func _initRoutes() {
 	e.Post("/syslog", querySyslog)
 
 	e.Get("/users", queryUsers)
+	e.Get("/users_skill/:id", queryUsersWithSkill)
 	e.Get("/users/:id", getUser)
 	e.Post("/users", newUser)
 	e.Put("/users/:id", saveUser)
@@ -40,6 +41,12 @@ func _initRoutes() {
 	e.Post("/sites", newSite)
 	e.Put("/sites/:id", saveSite)
 	e.Delete("/sites/:id", deleteSite)
+
+	e.Get("/skills", querySkills)
+	e.Get("/skills/:id", getSkill)
+	e.Post("/skills", newSkill)
+	e.Put("/skills/:id", saveSkill)
+	e.Delete("/skills/:id", deleteSkill)
 
 }
 
@@ -98,7 +105,6 @@ func sysLog(status int, t string, reftype string, ref int, descr string, c *echo
 	if claim != nil {
 		UserID, Username = getClaimedUser(claim)
 	}
-	log.Println("syslog", UserID, Username)
 
 	l := &DBsyslog{
 		Status:   status,
@@ -110,10 +116,15 @@ func sysLog(status int, t string, reftype string, ref int, descr string, c *echo
 		UserID:   UserID,
 		Username: Username,
 	}
-	DB.InsertInto("sys_log").
+
+	_, err := DB.InsertInto("sys_log").
 		Whitelist("status", "type", "ref_type", "ref_id", "ip", "descr", "user_id", "username").
 		Record(l).
 		Exec()
+
+	if err != nil {
+		log.Println("SysLog error", err.Error())
+	}
 }
 
 func querySyslog(c *echo.Context) error {
@@ -127,7 +138,6 @@ func querySyslog(c *echo.Context) error {
 	if err := c.Bind(req); err != nil {
 		return c.String(http.StatusBadRequest, err.Error())
 	}
-	log.Println("Syslog request", req)
 
 	if req.Limit < 20 {
 		req.Limit = 20
@@ -146,19 +156,20 @@ func querySyslog(c *echo.Context) error {
 		Limit(req.Limit)
 
 	// Add extra options to the SQL query
-	if req.RefType != "" {
-		query.Where("ref_type = $1", req.RefType)
-	}
-
-	if req.RefID != "" {
-		query.Where("ref_id = $1", req.RefID)
-	}
-
 	if req.UserID != "" {
+
 		// Grab any log records created by this specific user
-		query.Where("user_id = $1", req.UserID)
 		// And any log records of type U related to this specific user
-		query.Where("ref_type = 'U' and ref_id=$1", req.UserID)
+		query.Where("user_id = $1 or (ref_type = 'U' and ref_id=$1)", req.UserID, req.UserID)
+
+	} else {
+		if req.RefType != "" {
+			query.Where("ref_type = $1", req.RefType)
+		}
+
+		if req.RefID != "" {
+			query.Where("ref_id = $1", req.RefID)
+		}
 	}
 
 	var record []*DBsyslog
@@ -199,10 +210,8 @@ func login(c *echo.Context) error {
 	var res loginResponse
 	err = DB.
 		Select("u.id,u.username,u.name,u.role,u.site_id,s.name as sitename").
-		From(`
-			users u
-			left join site s on (s.id = u.site_id)
-		`).
+		From(`users u
+			left join site s on (s.id = u.site_id)`).
 		Where("u.username = $1 and passwd = $2", l.Username, l.Passwd).
 		QueryStruct(&res)
 
@@ -254,16 +263,18 @@ func logout(c *echo.Context) error {
 */
 
 type DBusers struct {
-	ID       int       `db:"id",json:"number"`
-	Username string    `db:"username"`
-	Passwd   string    `db:"passwd"`
-	Name     string    `db:"name"`
-	Email    string    `db:"email"`
-	Address  string    `db:"address"`
-	SMS      string    `db:"sms"`
-	SiteId   int       `db:"site_id"`
-	Role     string    `db:"role"`
-	Sites    []*DBsite `db:"sites"`
+	ID       int        `db:"id",json:"number"`
+	Username string     `db:"username"`
+	Passwd   string     `db:"passwd"`
+	Name     string     `db:"name"`
+	Email    string     `db:"email"`
+	Address  string     `db:"address"`
+	SMS      string     `db:"sms"`
+	SiteId   int        `db:"site_id"`
+	SiteName *string    `db."sitename"`
+	Role     string     `db:"role"`
+	Sites    []*DBsite  `db:"sites"`
+	Skills   []*DBskill `db:"skills"`
 }
 
 type DBuserlog struct {
@@ -281,6 +292,11 @@ type DBuserSite struct {
 	Role   string `db:"role"`
 }
 
+type DBuserSkill struct {
+	UserID  int `db:"user_id"`
+	SkillID int `db:"skill_id"`
+}
+
 func queryUsers(c *echo.Context) error {
 
 	_, err := securityCheck(c, "readUser")
@@ -291,7 +307,37 @@ func queryUsers(c *echo.Context) error {
 	var users []*DBusers
 	//		SQL(`select *,array(select concat(logdate,ip,descr) from user_log where user_id=users.id order by logdate desc) as logs from users`).
 
-	err = DB.SQL(`select * from users order by lower(username)`).QueryStructs(&users)
+	err = DB.SQL(`select 
+		users.*,site.name as sitename
+		from users 
+		left join site on site.id=users.site_id
+		order by lower(username)`).
+		QueryStructs(&users)
+
+	if err != nil {
+		return c.String(http.StatusNoContent, err.Error())
+	}
+	return c.JSON(http.StatusOK, users)
+}
+
+func queryUsersWithSkill(c *echo.Context) error {
+
+	_, err := securityCheck(c, "readUser")
+	if err != nil {
+		return c.String(http.StatusUnauthorized, err.Error())
+	}
+
+	var users []*DBusers
+	//		SQL(`select *,array(select concat(logdate,ip,descr) from user_log where user_id=users.id order by logdate desc) as logs from users`).
+
+	skillID := getID(c)
+	err = DB.SQL(`select 
+		u.*,t.name as sitename from user_skill s
+		left join users u on u.id = s.user_id
+		left join site t on t.id = u.site_id
+		where s.skill_id=$1
+		order by lower(username)`, skillID).
+		QueryStructs(&users)
 
 	if err != nil {
 		return c.String(http.StatusNoContent, err.Error())
@@ -313,6 +359,29 @@ func getUser(c *echo.Context) error {
 	if err != nil {
 		return c.String(http.StatusNoContent, err.Error())
 	}
+
+	// Fill in the sites for the user
+	DB.SQL(`select 
+		s.* 
+		from user_site u
+		left join site s on s.id=u.site_id 
+		where u.user_id=$1`, id).
+		QueryStructs(&user.Sites)
+
+	// Fill in the skills for the user
+	DB.SQL(`select 
+		s.* 
+		from user_skill u
+		left join skill s on s.id=u.skill_id 
+		where u.user_id=$1`, id).
+		QueryStructs(&user.Skills)
+
+	// Fill in the site name
+	DB.SQL(`select 
+		name as sitename from site 
+		where id=$1`, user.SiteId).
+		QueryScalar(&user.SiteName)
+
 	return c.JSON(http.StatusOK, user)
 }
 
@@ -328,8 +397,14 @@ func newUser(c *echo.Context) error {
 		return c.String(http.StatusBadRequest, err.Error())
 	}
 
+	if len(record.Sites) > 0 {
+		record.SiteId = record.Sites[0].ID
+	} else {
+		record.SiteId = 0
+	}
+
 	err = DB.InsertInto("users").
-		Whitelist("username", "passwd", "address", "name", "email", "role", "sms").
+		Whitelist("username", "passwd", "address", "name", "email", "role", "sms", "site_id").
 		Record(record).
 		Returning("id").
 		QueryScalar(&record.ID)
@@ -338,11 +413,39 @@ func newUser(c *echo.Context) error {
 		return c.String(http.StatusInternalServerError, err.Error())
 	}
 
+	// Now create the user_site records
+	for _, addSite := range record.Sites {
+		userSite := &DBuserSite{
+			UserID: record.ID,
+			SiteID: addSite.ID,
+			Role:   record.Role,
+		}
+		_, xe := DB.InsertInto("user_site").
+			Whitelist("user_id", "site_id", "role").
+			Record(userSite).
+			Exec()
+		if xe != nil {
+			log.Println("ERROR inserting user_site", xe.Error())
+		}
+	}
+
+	// Now create the user_skill recordsd
+	for _, addSkill := range record.Skills {
+		userSkill := &DBuserSkill{
+			UserID:  record.ID,
+			SkillID: addSkill.ID,
+		}
+		_, xe := DB.InsertInto("user_skill").Whitelist("user_id", "skill_id").Record(userSkill).Exec()
+		if xe != nil {
+			log.Println("ERROR inserting user_skill", xe.Error())
+		}
+	}
+
 	// Now log the creation of the new user
 	sysLog(1, "Users", "U", record.ID, fmt.Sprintf("Account Created - %s", record.Username), c, claim)
 
 	// insert into DB, fill in the ID of the new user
-	return c.JSON(http.StatusCreated, newUser)
+	return c.JSON(http.StatusCreated, record)
 }
 
 func saveUser(c *echo.Context) error {
@@ -358,6 +461,7 @@ func saveUser(c *echo.Context) error {
 	}
 
 	userID := getID(c)
+	log.Println("saveUser", record, userID)
 
 	// Get the first site from the list of sites added
 	if len(record.Sites) > 0 {
@@ -366,24 +470,29 @@ func saveUser(c *echo.Context) error {
 		record.SiteId = 0
 	}
 
+	log.Println("record2", record)
+	log.Println("record3.1", record.SiteName)
 	_, err = DB.Update("users").
-		SetWhitelist(record, "username", "name", "passwd", "email", "address", "sms", "role", "site_id").
+		Set("username", record.Username).
+		Set("name", record.Name).
+		Set("passwd", record.Passwd).
+		Set("email", record.Email).
+		Set("address", record.Address).
+		Set("sms", record.SMS).
+		Set("role", record.Role).
+		Set("site_id", record.SiteId).
 		Where("id = $1", userID).
 		Exec()
+	log.Println("record3", record, err)
 
 	if err != nil {
 		return c.String(http.StatusInternalServerError, err.Error())
 	}
 
 	// Now create the user_site records
-	DB.DeleteFrom("user_site").
-		Where("user_id=$1", userID).
-		Exec()
+	DB.DeleteFrom("user_site").Where("user_id=$1", userID).Exec()
 
-	log.Println("Sites = ", record.Sites)
 	for _, addSite := range record.Sites {
-		log.Println("Attach user", userID, "to site", addSite.ID, "name", addSite.Name)
-
 		userSite := &DBuserSite{
 			UserID: userID,
 			SiteID: addSite.ID,
@@ -393,6 +502,16 @@ func saveUser(c *echo.Context) error {
 			Whitelist("user_id", "site_id", "role").
 			Record(userSite).
 			Exec()
+	}
+
+	// Now create the user_skill recordsd
+	DB.DeleteFrom("user_skill").Where("user_id=$1", userID).Exec()
+	for _, addSkill := range record.Skills {
+		userSkill := &DBuserSkill{
+			UserID:  userID,
+			SkillID: addSkill.ID,
+		}
+		DB.InsertInto("user_skill").Whitelist("user_id", "skill_id").Record(userSkill).Exec()
 	}
 
 	sysLog(1, "Users", "U", userID, "User Updated", c, claim)
@@ -523,22 +642,30 @@ func deleteSite(c *echo.Context) error {
 	return c.String(http.StatusOK, "TODO delete site")
 }
 
-func querySitelog(c *echo.Context) error {
+///////////////////////////////////////////////////////////////////////
+// Skills Maintenance
+/*
+create table skill (
+	id serial not null primary key,
+	name text not null
+);
 
-	_, err := securityCheck(c, "readSite")
+*/
+
+type DBskill struct {
+	ID   int    `db:"id"`
+	Name string `db:"name"`
+}
+
+func querySkills(c *echo.Context) error {
+
+	_, err := securityCheck(c, "readSkill")
 	if err != nil {
 		return c.String(http.StatusUnauthorized, err.Error())
 	}
 
-	id := getID(c)
-	var record []*DBuserlog
-	err = DB.SQL(`
-		select type,ref_id,to_char(logdate,'Dy DD-Mon-YY HH24:MI:SS') as logdate,ip,descr,username
-		from sys_log l
-		where ref=$1 and ref_type='S' 
-		order by l.logdate desc
-		limit 20`, id).
-		QueryStructs(&record)
+	var record []*DBskill
+	err = DB.SQL(`select * from skill order by lower(name)`).QueryStructs(&record)
 
 	if err != nil {
 		return c.String(http.StatusNoContent, err.Error())
@@ -546,24 +673,80 @@ func querySitelog(c *echo.Context) error {
 	return c.JSON(http.StatusOK, record)
 }
 
-func querySitelogs(c *echo.Context) error {
+func getSkill(c *echo.Context) error {
 
-	_, err := securityCheck(c, "readSite")
+	_, err := securityCheck(c, "readSkill")
 	if err != nil {
 		return c.String(http.StatusUnauthorized, err.Error())
 	}
 
-	var record []*DBuserlog
-	err = DB.SQL(`
-		select type,ref_id,to_char(logdate,'Dy DD-Mon-YY HH24:MI:SS') as logdate,ip,descr,username
-		from sys_log l 
-		where ref_type='S' 
-		order by l.logdate desc
-		limit 50`).
-		QueryStructs(&record)
+	id := getID(c)
+	var record DBskill
+	err = DB.SQL(`select * from skill where id=$1`, id).QueryStruct(&record)
 
 	if err != nil {
 		return c.String(http.StatusNoContent, err.Error())
 	}
 	return c.JSON(http.StatusOK, record)
+}
+
+func newSkill(c *echo.Context) error {
+
+	claim, err := securityCheck(c, "writeSkill")
+	if err != nil {
+		return c.String(http.StatusUnauthorized, err.Error())
+	}
+
+	record := &DBskill{}
+	if err := c.Bind(record); err != nil {
+		return c.String(http.StatusBadRequest, err.Error())
+	}
+
+	err = DB.InsertInto("skill").
+		Whitelist("name").
+		Record(record).
+		Returning("id").
+		QueryScalar(&record.ID)
+
+	if err != nil {
+		return c.String(http.StatusInternalServerError, err.Error())
+	}
+
+	// Now log the creation of the new site
+	sysLog(1, "Skills", "s", record.ID, "Skill Created", c, claim)
+
+	// insert into DB, fill in the ID of the new user
+	return c.JSON(http.StatusCreated, record)
+}
+
+func saveSkill(c *echo.Context) error {
+
+	claim, err := securityCheck(c, "writeSkill")
+	if err != nil {
+		return c.String(http.StatusUnauthorized, err.Error())
+	}
+
+	record := &DBskill{}
+	if err = c.Bind(record); err != nil {
+		return c.String(http.StatusBadRequest, err.Error())
+	}
+
+	skillID := getID(c)
+
+	_, err = DB.Update("site").
+		SetWhitelist(record, "name").
+		Where("id = $1", skillID).
+		Exec()
+
+	if err != nil {
+		return c.String(http.StatusInternalServerError, err.Error())
+	}
+
+	sysLog(1, "Skills", "s", skillID, "Updated", c, claim)
+	return c.JSON(http.StatusOK, skillID)
+}
+
+func deleteSkill(c *echo.Context) error {
+
+	return c.String(http.StatusOK, "TODO delete skill")
 }
